plugins/vim: fix socket.send() error
When the connection is lost, the transport isn't being removed from the
set of clients, so when it's time to send an update, the update is sent
to all connections including closed ones.  This change makes sure that
the set of transports includes only connected transports by removing
closed transports via the `connection_lost` callback.