color: add generate_from classmethod
This is borrowed from "gimme_colors" in my (private) larry-command repo.
Given a list of colors, generate the given number of needed colors in a
semi-smart way.

So, e.g. If I have 2 colors and need 12, it will generate a gradient
from darkest to lightest.  If I have 12 colors and need 2, it will
generate 2 random colors from the list of 12, etc.