gnome-shell: create new theme dir
Instead of editing gnome-shell.css in a current theme and reloading the
theme (which doesn't always work). Create a new theme (dir) using a
with a random name, write the new gnome-shell.css there and then switch
to that theme.

Remove the previous theme if it was created by larry.