color: replace replace_string with replace_string2
As I was writing tests for replace_string and it's helper functions I
thought "this is just crazy". I don't like the way it works at all so
decided to replace it with replace_string2. The gtk and gnome-shell
plugins have been changed to use replace_string2, but later I'll gut
them entirely too to not depend on the input colors being in the config.

I've left replace_string around until I've determined that nothing else
is using it.