plugins: refactor gnome_terminal plugin
I feel bad after commit c38086c and want to reduce the number of @mocks
in the test and I also think that the theme as a class has some
advantages. This commit addresses both.

Note that "theme" now means the entire directory under ~/.themes and not
just the gnome-shell subdirectory. As such the plugin's template setting
has a different meaning but is more self-explanatory.