plugins: remove the filtered decorator
The filtered decorator was intended to work similarly to what
`apply_filter_plugin` does now, but it applied to the plugin itself and
the plugin only sees the filtered colors. In reality it didn't make
sense which is the reason why it was never used. This change removes the
decorator and its helper function.